test(integration): share redis oauth fixtures
